Исправлены комментарии после ПР
