+++ FakeNFT/Helpers/ViewCornerRadiusExtention.swift
@@ -1,0 +1,29 @@
+//
+//  ViewCornerRadiusExtention.swift
+//  FakeNFT
+//
+//  Created by Maksim Zakharov on 03.06.2025.
+//
+
+import SwiftUI
+
+// MARK: - Helper Extension
+extension View {
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape(RoundedCorner(radius: radius, corners: corners))
+    }
+}
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(
+            roundedRect: rect,
+            byRoundingCorners: corners,
+            cornerRadii: CGSize(width: radius, height: radius)
+        )
+        return Path(path.cgPath)
+    }
+}
